Fuji: - Updated some libraries. - Minor 64bit portability problems... 64bit builds again. - Fixed some of the resource pools, improved speed and memory usage.
Haku:
- Added a Widget styling system.

git-svn-id: https://fuji.googlecode.com/svn/trunk@815 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Source/MFFileSystem.cpp
@@ -416,14 +416,14 @@ int MFFile_Close(MFFile* fileHandle)
 }
 
 // read/write functions
-int MFFile_Read(MFFile* fileHandle, void *pBuffer, uint32 bytes, bool async)
+int MFFile_Read(MFFile* fileHandle, void *pBuffer, size_t bytes, bool async)
 {
 	MFCALLSTACK;
 
 	return ppFileSystemList[fileHandle->filesystem]->callbacks.Read(fileHandle, pBuffer, bytes);
 }
 
-int MFFile_Write(MFFile* fileHandle, const void *pBuffer, uint32 bytes, bool async)
+int MFFile_Write(MFFile* fileHandle, const void *pBuffer, size_t bytes, bool async)
 {
 	MFCALLSTACK;
